# doc/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Library Backend API",
    "version": "1.0.0",
    "description": "API para gerenciamento de usuarios, colecaes e livros favoritos com autenticacao JWT"
  },
  "host": "127.0.0.1:5000",
  "basePath": "/",
  "schemes": ["http"],

  "tags": [
    {
      "name": "Auth",
      "description": "Autenticacaoo de usuarios"
    },
    {
      "name": "Colecoes",
      "description": "Gerenciamento de colecoes"
    },
    {
      "name": "Livros",
      "description": "Gerenciamento de livros"
    }
  ],

  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "autorizacao",
      "in": "header",
      "description": "Cole o {seu_token}"
    }
  },

  "definitions": {
    "Usuario": {
      "type": "object",
      "properties": {
        "nome": { "type": "string" },
        "email": { "type": "string" },
        "senha": { "type": "string" }
      }
    },
    "Login": {
      "type": "object",
      "properties": {
        "email": { "type": "string" },
        "senha": { "type": "string" }
      }
    },
    "Livro": {
      "type": "object",
      "properties": {
        "titulo": { "type": "string" },
        "ano": { "type": "integer" },
        "descricao": { "type": "string" },
        "autor": { "type": "string" },
        "capa": { "type": "string" }
      }
    },
    "Colecao": {
      "type": "object",
      "properties": {
        "nome": { "type": "string" }     
      }
    }
    },


  "paths": {
    "/login": {
    "post": {
      "tags": ["Auth"],
      "summary": "Autenticar usuario",
      "consumes": ["application/json"],
      "produces": ["application/json"],
      "parameters": [
        {
          "name": "body",
          "in": "body",
          "required": true,
          "schema": { "$ref": "#/definitions/Login" }
        }
      ],
      "responses": {
        "200": {
          "description": "Login realizado",
          "schema": {
            "type": "object",
            "properties": {
              "status": { "type": "string" },
              "usuario_id": { "type": "integer" },
              "nome": { "type": "string" },
              "autorizacao": { "type": "string" }}
            
          }
        }
      }
    }
    },

    "/cadastro": {
      "post": {
        "tags": ["Auth"],
        "summary": "Cadastrar usuario",
        "produces": [
          "application/json"
          ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/Usuario" }
          }
        ],
        "responses": {
          "201": { "description": "Usuario criado" },
          "400": { "description": "Erro de validacao" }
        }
      }
    },

    "/mostrar_colecao": {
      "get": {
        "tags": ["Colecoes"],
        "summary": "Listar colecaes do usuario",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Lista de colecoes" },
          "401": { "description": "Nao autorizado" }
        }
      }
    },

    "/nova_colecao": {
      "post": {
        "tags": ["Colecoes"],
        "summary": "Criar nova colecao",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nome": { "type": "string" }
              }
            }
          }
        ],
        "responses": {
          "201": { "description": "Colecao criada" }
        }
      }
    },

    "/deletar_colecao/{colection_id}": {
      "delete": {
        "tags": ["Colecoes"],
        "summary": "Remover colecao",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "colection_id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": { "description": "Coleaoo removida" },
          "404": { "description": "Nao encontrada" }
        }
      }
    },

    "/add_livro/{colection_id}": {
      "post": {
        "tags": ["Livros"],
        "summary": "Adicionar livro aos favoritos",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "colection_id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/Livro" }
          }
        ],
        "responses": {
          "201": { "description": "Livro adicionado" },
          "403": { "description": "Não autorizado" },
          "404": { "description": "Coleção não encontrada" }
        }
      }
    },

    "/mostrar_livros/{colection_id}": {
      "get": {
        "tags": ["Livros"],
        "summary": "Listar livros favoritos de uma coleção",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "colection_id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": { "description": "Lista de livros" },
          "403": { "description": "Não autorizado" },
          "404": { "description": "Coleção não encontrada" }
        }
      }
    },

    "/deletar_livro/{livro_id}": {
      "delete": {
        "tags": ["Livros"],
        "summary": "Remover livro dos favoritos",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "livro_id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": { "description": "Livro removido" },
          "404": { "description": "Livro não encontrado" }
        }
      }
    }
    
  }
}